pyproject.toml
[project]
name = "uv-trial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "faker>=33.3.1",
    "findspark>=2.0.1",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.19.0",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.1.0",
    "pathlib>=1.0.1",
    "pyspark>=3.5.4",
    "python-dotenv>=1.0.1",
]
